requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
filelock==3.18.0
geographiclib==2.0
geopy==2.4.1
h3==4.2.2
idna==3.10
image2pdf==1.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
img2pdf==0.6.1
lxml==5.4.0
moviepy==2.2.1
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pikepdf==9.8.1
pillow==11.2.1
platformdirs==4.3.8
proglog==0.1.12
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
timezonefinder==6.5.9
tkinterdnd2==0.4.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
wrapt==1.17.2
